+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - from_dict method to HazardConfig and OpenquakeConfig classes
 - factory method for obtaining HazardConfig type
+- z1.0 and z2.5 calculation function ported from openquake
 
 ### Changed
 - hazard config for model versions stored as file